IoT/air_quality/container/requirements.txt
pyserial
requests
paho-mqtt==1.6.1
entrypoints
executing